+++ packages/start-plugin-core/src/start-compiler/compiler.ts
@@ -1451,7 +1451,7 @@ export class StartCompiler {
         parseExpression: (expressionCode) =>
           babel.template.expression(expressionCode, {
             placeholderPattern: false,
-          })() as t.Expression,
+          })(),
         warn: warnFn,
 
         generateFunctionId: (opts) => this.generateFunctionId(opts),
@@ -1580,7 +1580,7 @@ export class StartCompiler {
         parseExpression: (expressionCode: string) =>
           babel.template.expression(expressionCode, {
             placeholderPattern: false,
-          })() as t.Expression,
+          })(),
         warn,
       }