+++ packages/solid-start/CHANGELOG.md
@@ -1,5 +1,92 @@
 # @tanstack/solid-start
 
+## 1.168.55
+
+### Patch Changes
+
+- Updated dependencies [[`222300b`](https://github.com/TanStack/router/commit/222300b82fe6f2a5e7e7f460ae21a2e35c6230d7)]:
+  - @tanstack/start-plugin-core@1.171.47
+
+## 1.168.54
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/solid-router@1.170.36
+  - @tanstack/solid-start-client@1.168.35
+  - @tanstack/solid-start-server@1.167.42
+  - @tanstack/start-client-core@1.170.32
+  - @tanstack/start-plugin-core@1.171.46
+  - @tanstack/start-server-core@1.169.37
+
+## 1.168.53
+
+### Patch Changes
+
+- Updated dependencies [[`e561fa1`](https://github.com/TanStack/router/commit/e561fa1d7118e3d29267cc3b6ce1130d6581f387), [`cbbfbe3`](https://github.com/TanStack/router/commit/cbbfbe37ab1dbe328c343cb437c5660769cc9f26), [`a1c8d1a`](https://github.com/TanStack/router/commit/a1c8d1aa759c227eae9601a030321ac4c53c24bd), [`cbbfbe3`](https://github.com/TanStack/router/commit/cbbfbe37ab1dbe328c343cb437c5660769cc9f26)]:
+  - @tanstack/solid-router@1.170.35
+  - @tanstack/start-plugin-core@1.171.45
+  - @tanstack/solid-start-server@1.167.41
+  - @tanstack/start-client-core@1.170.31
+  - @tanstack/start-server-core@1.169.36
+  - @tanstack/solid-start-client@1.168.34
+
+## 1.168.52
+
+### Patch Changes
+
+- [#8397](https://github.com/TanStack/router/pull/8397) [`ab99818`](https://github.com/TanStack/router/commit/ab99818225835a1553a9f3513cd2093ff1b257be) - Support Rsbuild 2.2: preserve function-based Rspack watch exclusions and regular-expression flags when adding monorepo build-output exclusions, and leave string-valued Babel loader options intact in the Solid Start plugin.
+
+- Updated dependencies [[`ab99818`](https://github.com/TanStack/router/commit/ab99818225835a1553a9f3513cd2093ff1b257be)]:
+  - @tanstack/start-plugin-core@1.171.44
+
+## 1.168.51
+
+### Patch Changes
+
+- Updated dependencies [[`9b2adaf`](https://github.com/TanStack/router/commit/9b2adaf8c3fca905f156c42f6a4fe17a787afcc8), [`9b2adaf`](https://github.com/TanStack/router/commit/9b2adaf8c3fca905f156c42f6a4fe17a787afcc8), [`634da91`](https://github.com/TanStack/router/commit/634da9176e16fa8aa49af459bbb054fcae7d85ef), [`e9396c9`](https://github.com/TanStack/router/commit/e9396c928945d1dd6fd3f3bd8052143794f688b5), [`9872d2a`](https://github.com/TanStack/router/commit/9872d2ac39fc05f4ef6566c0b421f71ceb115244), [`9b2adaf`](https://github.com/TanStack/router/commit/9b2adaf8c3fca905f156c42f6a4fe17a787afcc8)]:
+  - @tanstack/solid-router@1.170.34
+  - @tanstack/start-server-core@1.169.35
+  - @tanstack/solid-start-client@1.168.33
+  - @tanstack/solid-start-server@1.167.40
+  - @tanstack/start-client-core@1.170.30
+  - @tanstack/start-plugin-core@1.171.43
+
+## 1.168.50
+
+### Patch Changes
+
+- Updated dependencies [[`b1c9219`](https://github.com/TanStack/router/commit/b1c92194397e57cecfab72de70e2dd8064a4f6d5), [`76239a3`](https://github.com/TanStack/router/commit/76239a34b732ca19d1b16d42efb83824b1de1c69), [`d914b08`](https://github.com/TanStack/router/commit/d914b085bb4bf29ec18a531201e377c10681b661)]:
+  - @tanstack/solid-router@1.170.33
+  - @tanstack/solid-start-client@1.168.32
+  - @tanstack/start-server-core@1.169.34
+  - @tanstack/solid-start-server@1.167.39
+  - @tanstack/start-plugin-core@1.171.42
+
+## 1.168.49
+
+### Patch Changes
+
+- Updated dependencies [[`49bcd4d`](https://github.com/TanStack/router/commit/49bcd4db743dba4a783d5533a42a0f97a5721292), [`9c1871c`](https://github.com/TanStack/router/commit/9c1871ccc88bc1157186b8b460941a304ec740b8), [`919c397`](https://github.com/TanStack/router/commit/919c3971493374cccc963e75b9b2a4e98ebf878a)]:
+  - @tanstack/start-plugin-core@1.171.41
+  - @tanstack/solid-router@1.170.32
+  - @tanstack/start-server-core@1.169.33
+  - @tanstack/solid-start-client@1.168.31
+  - @tanstack/solid-start-server@1.167.38
+  - @tanstack/start-client-core@1.170.29
+
+## 1.168.48
+
+### Patch Changes
+
+- Updated dependencies [[`2f20c00`](https://github.com/TanStack/router/commit/2f20c00224c5ba63467551914e0c37012588c4c2), [`28a5e45`](https://github.com/TanStack/router/commit/28a5e4504e4ea5cb1480667a4bea2588a53e110f), [`f0b5eda`](https://github.com/TanStack/router/commit/f0b5eda544606686a8a8d675a686ca1366428b96), [`37877da`](https://github.com/TanStack/router/commit/37877da166fe4ce055c7b85e138b6681ebd7e8b4), [`0497cae`](https://github.com/TanStack/router/commit/0497caeef3ff7e1c1c6080eca38bca24e7ec320b), [`0caf6b9`](https://github.com/TanStack/router/commit/0caf6b9a2b7e14b0b146c74cc27cb05c19d700a5)]:
+  - @tanstack/solid-router@1.170.31
+  - @tanstack/start-plugin-core@1.171.40
+  - @tanstack/solid-start-client@1.168.30
+  - @tanstack/solid-start-server@1.167.37
+  - @tanstack/start-client-core@1.170.28
+  - @tanstack/start-server-core@1.169.32
+
 ## 1.168.47
 
 ### Patch Changes